requirements.txt
matplotlib==3.5.3
numpy==1.22.4
pandas==1.4.2
Pillow==9.2.0
plotly==5.10.0
requests==2.27.1
scipy==1.9.1
seaborn==0.11.2
streamlit==1.9.2
streamlit_lottie==0.0.3
